+++ components/eventForm/toolbar/index.tsx
@@ -1,19 +1,10 @@
-import React, {
-  createContext,
-  ReactNode,
-  useContext,
-  useRef,
-  useState
-} from "react"
+import React from "react"
 import { Button, Text } from "react-native"
-import { useEventFormValue } from "./EventForm"
+import { useEventFormValue } from "../EventForm"
 import { TouchableOpacity } from "react-native-gesture-handler"
 import { HexColorPickerOption } from "@components/formComponents/HexColorPicker"
 import { EventColors } from "@lib/events/EventColors"
-import {
-  BottomSheetModal,
-  BottomSheetModalProvider
-} from "@gorhom/bottom-sheet"
+import { ToolbarProvider, useToolbar } from "./ToolbarProvider"
 
 /**
  * A horizontally scrolling toolbar for an event form.
@@ -44,17 +35,9 @@ const DateTab = () => {
   )
 }
 
-const DateSection = () => {
-  return <Text>Start and End Dates</Text>
-}
-
 const ColorTab = () => {
   const { openSection } = useToolbar()
   return <Button title="Color" onPress={() => openSection("color")} />
-}
-
-const ColorSection = () => {
-  return <Text>Pick Color</Text>
 }
 
 const eventColorOptions = [
@@ -79,62 +62,5 @@ const AdvancedTab = () => {
       accessibilityLabel="Advanced"
       onPress={() => openSection("advanced")}
     />
-  )
-}
-
-const AdvancedSection = () => {
-  return <Text>Advanced</Text>
-}
-
-type ToolbarSection = "date" | "color" | "advanced"
-
-type ToolbarContextValues = {
-  openSection: (section: ToolbarSection) => void
-}
-
-const ToolbarContext = createContext<ToolbarContextValues | undefined>(
-  undefined
-)
-
-const useToolbar = () => useContext(ToolbarContext)!!
-
-type ToolbarProviderProps = {
-  children: ReactNode
-}
-
-const bottomSheetSnapPoints = ["50%"]
-
-const ToolbarProvider = ({ children }: ToolbarProviderProps) => {
-  const bottomSheetRef = useRef<BottomSheetModal>(null)
-  const [section, setSection] = useState<ToolbarSection | undefined>()
-
-  return (
-    <ToolbarContext.Provider
-      value={{
-        openSection: (section) => {
-          setSection(section)
-          bottomSheetRef.current?.present()
-        }
-      }}
-    >
-      <BottomSheetModalProvider>
-        <BottomSheetModal
-          ref={bottomSheetRef}
-          snapPoints={bottomSheetSnapPoints}
-        >
-          <TouchableOpacity
-            accessibilityLabel="Close Section"
-            onPress={() => {
-              setSection(undefined)
-              bottomSheetRef.current?.dismiss()
-            }}
-          />
-          {section === "date" && <DateSection />}
-          {section === "color" && <ColorSection />}
-          {section === "advanced" && <AdvancedSection />}
-        </BottomSheetModal>
-      </BottomSheetModalProvider>
-      {children}
-    </ToolbarContext.Provider>
   )
 }
